Don't reset timestamps on app context files
In an attempt to ensure fully reproducible builds, when we built the
temp file context for an app build, we were setting the timestamp of
all the files to 0. This caused an issue where transferring the context
to buildkit would not pickup changes in some cases. This is because
that is done based on metadata only, not content of the files. So, a
file change that result in the same file size wouldn't get transferred.

This buildkit issue comment explains the situation even better:
https://github.com/moby/buildkit/issues/2592#issuecomment-1027642738