Drop the autogen tag from docs
We have a CI check that ensures the generated docs are up-to-date. Cobra
appends a line at the bottom of the docs that says the date the docs
were generated on. So, that change in the docs was causing our CI to
fail.

This fixes that by eliminating the auto-gen tag.