Control AcornDNS default config from build
Set the AcornDNS endpoint from the build so that "real" releases are
pointed at the production endpoint and builds from main or done locally
are pointing at the staging endpoint.

This also sets the default state as disabled. as build arguement. Once
the AcornDNS service is up and running, we'll change the default to auto.