Add label and annotation support
This change adds support for adding labels and annotations to the
resources created by acorn. The labels and annotations can be specified
in the Acornfile or on the command line and can be scoped to apply to
all resources, resources of a specific type (ie containers or volumes),
or resources with a specific name.

Signed-off-by: Craig Jellick <craig@acorn.io>