Support numbers with G suffix in AML
The size of a volume was assumed to be a string but numbers were
supported also.  If a number was put in like 5 it was assumed to be
gigabytes.  The issue happens where in the field you put 5G instead
of "5G" (quoted). The 5G syntax is parsed by AML and turned into
5000000000.  Then we assume that is gigabytes turning into
5000000000G.

The fix is a bit of a hack where we basically say if the number is less
than 1000000 we treat it as gigabytes.

Signed-off-by: Darren Shepherd <darren@acorn.io>