Set ingress class name when we need it
A common scenario we see is that even though ingress-nginx is installed,
it isn't set as the default ingress class. In this case, when we create an
ingress, it never gets picked up.

So, to make ingress work in this scenario, if we detect that the cluster
has exactly one ingressClass and it isn't set at the default, we are
just going to assume the user wants to use it and set it when we create
an ingress.

Bonus: found a bug where if Docker Desktop has an ingress already
installed, we still set acorn's ingress class name to "traefik". We
should only set the ingressClassName field for Docker Desktop if we are
installing traefik.

Signed-off-by: Craig Jellick <craig@acorn.io>