Pass params to validation of permissions and update launch.json
Prior to this commit, validation of permissions was not being calculated
using the params set by the users. For example, profiles were not being
taken into account when setting conditional permissions. On top of fixing
this, this also adds an update to launch.json to allow specifying the
Acornfile or image to run.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>