Add ability to set workload memory via limit/request (#43)
This commit adds three ways to set the limit/request restriction on memory. In
order or precedence, they are:

1. "acorn run --memory workload=512Mi" where workload is the name of a
   workload in the Acornfile and 512Mi is 512 mebibytes.
2. A new property in the Acornfile for defined containers - memory. It
   takes a Cuelang integer that equates to bytes.
3. "acorn install --workload-memory-default 512Mi" where 512Mi is the
   memory that Acorn will default to if no other option is set.

These flags all get evaluated by the API Server and Controller to prevent
any invalid memory values being set. This commit also adds the
`--workload-memory-maximum` flag which will prevent an acorn exceeding
the limit to install. These new flags of course bring changes to the
documentation and tests to fight off regressions.

Lastly, there is a new type of "builder" flag, "quantity", that can be
used to turn string values from the user into int64's on the backend.

Co-authored-by: Joshua Silverio <joshua@acorn.io>
Signed-off-by: tylerslaton <mtslaton1@gmail.com>
Signed-off-by: Joshua Silverio <joshua@acorn.io>