Add CLI completion for apps and containers
These changes add support for completion of app names and container
names, both acorn containers and container replicas, in the CLI. A few
of these are very straightforward: apps, start, stop, update, wait, etc.
Others are more complicated and their "why" is described below. However,
to avoid code and test duplication, the completion logic is separated
from the command logic. Then, each command that needs completion, can
use this logic without duplicating it.

The more complication completions are for things like logs and exec. For
each of these commands, a user can provide a container replica name or
an app name as the argument. The functionality for completion is app
name completion until a '.' is found in the string and then container
replica name completion. This is allowed because of the convention that
container replica names are of the form
`<APP_NAME>.<ACORN_CONTAINER>-<RANDOM>`.

The most complicated is the `-c` flag for exec. If the `-c` flag is not
provided, then the functionality is described as above. The completion
for the `-c` flag is for containers as defined in the Acornfile. If the
`-c` is given and the user is completing the argument, then only apps
with the given container name are given in the completion.

Signed-off-by: Donnie Adams <donnie@acorn.io>