Add CLI completion for images
We always prefer image tags over image digests. That is, if an image
digest matches a completion, but that image has a tag, then the tag will
complete and the image digest will not.

Secondly, there were two log messages that were incorrect when testing
some of this, so I fixed them here as well.

Signed-off-by: Donnie Adams <donnie@acorn.io>