Avoid object lookups in config if acorn dns is disabled
Currently we always lookup all v1.Node objects to determine if we should
lookup and use an Acorn DNS provisioned domain. We do this even if
Acorn DNS is disabled, and other conditions where the v1.Node lookup is
not required. This change makes it such that if Acorn DNS is specifically
disabled then don't lookup v1.Node objects.

Signed-off-by: Darren Shepherd <darren@acorn.io>