Refactor the auto-upgrade logic for simplicity and testability
The auto-upgrade sync mechanism had three ways of being started: direct
call to Sync, the ticker based on the cluster-wide auto upgrade
interval, and a sleep in the sync function that would kick off the next
sync when the next app is due to be updated or when the next
cluster-wide auto upgrade interval occurs, whichever is first. Since the
second and third scenarios both take the cluster-wide configuration into
account, this has been simplified into one timer mechanism.

Additionally, there was an edge-case that was not accounted for in the
auto upgrade logic: deleting and recreating an app that used the
auto-upgrade functionality within a sync duration of this logic. The
sync would be kicked off by the detection of an app that uses
auto-upgrade, but the upgrade logic would not detect that this is a new
app and not the previously deleted one. The logic would dictate that the
app image was not in need of an upgrade and progress to run the app
would hang. This is also addressed by these changes.

Unit tests have also been added to test this functionality.

These changes made it possible for the ImageDetails API to accept and
handle auto-upgrade tag patterns. The immediate benefit of this is to
allow validation of acorn args when the user passes an auto-upgrade tag
to `acorn run`.

Signed-off-by: Donnie Adams <donnie@acorn.io>