Fix auto-upgrade with local images
If auto-upgrade is used, then only remote images will be checked. If a
user tries to run an app with auto-upgrade enabled and a local image not
pushed to a remote repo, then the apps won't start.

After these changes, any images or tags in the local repo are used for
auto-upgrades.

Signed-off-by: Donnie Adams <donnie@acorn.io>