Use ProjectList to test valid clients
Before Info was used to test if a client is valid, but that assumes some
namespace exists.  ProjectList is cluster scoped and makes no other
assumptions on the state of the cluster.  Also due to Project list looking
at namespace privileges, all users should be granted the ability to list
projects.

Signed-off-by: Darren Shepherd <darren@acorn.io>