Support allow-list for user labels and annotations
A previous change made it possible to disable propagation of
user-defined labels and annotations. This turned out to be too
restrictive, and we need a way to allow only certain labels and
annotations.

This change adds two installation flags: allow-user-labels and
allow-user-annotations. If values are supplied for these flags and the
ignore-user-labels-and-annotations is true, then only the allowed labels
and annotations are propagated. If ignore-user-labels-and-annotations is
not set or is false, then these new flags have no effect.

Signed-off-by: Donnie Adams <donnie@acorn.io>