Add recorded field to acornimagebuildinstance
The build server saves the ImageInstance to acorn.  In a hub context
that is written to the downstream server which them must be synced up
to the account api server. This recorded field is here to enable hub to
notify the build server when the image has been synced and the build can
return.

Signed-off-by: Darren Shepherd <darren@acorn.io>