Don't unset the env vars
If you call the DefaultClient() method twice from the client factory what
happens is we read the ENV vars, then lookup unsets them, and then on
the second call it reads the env vars and it's empty.

It was obviously a bad idea to unset the env vars in the first place, that
was just a bad idea on my part.  So we just won't do that anymore.

Signed-off-by: Darren Shepherd <darren@acorn.io>