Fix inability to run `acorn run -i` on app with missing image
If the app.spec.image is pointing to an invalid image (one that
can not be pulled for whatever reason) all updates to the app will
fail by design.  The issue is that in `acorn run -i .` we try to
update the app to running before we update the app to a new valid
image that was built. This PR fixes the behavior so that we will
always first update the image before making any other changes to the
app.

Signed-off-by: Darren Shepherd <darren@acorn.io>