Add PVC status to CLI
By default, the CLI will show the status of different objects in an app
until the app is ready. The status of volumes/PVCs is not shown. If
there is an issue, the status will only show information about the
container.

With this change, the CLI will show status for PVCs as volumes.

Signed-off-by: Donnie Adams <donnie@acorn.io>