Fix pull/run with internal-registry-prefix
If internal-registry-prefix is used then doing pull of an image
and subsequent run of that image will fail. The actual external
location of the repo was not getting recorded in the same fashion
that the buildserver does.  What is supposed to happen is the Repo
field of the ImageInstance should be set to the fully qualified reference
in the external repo.

Signed-off-by: Darren Shepherd <darren@acorn.io>