Don't decrypted data when writing generated secret
If the user puts an encrypted value a secret data field
in the Acornfile do not decrypt the value when creating the generated
secret in the app's source namespace. This causes an undesired side
effect  in that the user can now do "acorn secret reveal" to get the
clear text of an encrypted secret. This isn't a secret issue because
the users has already been granted "reveal", but we'd rather not
reveal the clear text value for a secret that was intentionally
encrypted by the user.

Signed-off-by: Darren Shepherd <darren@acorn.io>