Add VolumeClass to the API and CLI
VolumeClasses are an abstraction of Storage Classes in Kubernetes. An
admin can create ProjectVolumeClasses and ClusterVolumeClasses to
indicate which Storage Classes are available to users, volume size
constraints, access mode constraints, and defaults if a user omits
information. If a user attempts to use a Volume Class that does not
exist, an error is produced.

The defaults for an app are calculated upfront. These defaults will not
change unless the generation of the app is updated. This allows running
apps to continue to run after an admin updates their Acorn installation.

Finally, there is a new API for users for VolumeClasses. This API does
not expose which VolumeClasses are Cluster or Project scoped. If a user
were to use `acorn offerings volumeclasses`, then they would be shown
all classes available to them.

An installation flag has been added that specifies that an admin would
like to manually manage the volume classes. If this flag is enabled,
then all volume classes would have to be created by an admin before any
existing storage classes can be used.

If this flag is not passed, then a controller will sync all existing
storage classes (and the ephemeral class) to ClusterVolumeClasses so
that users can use these classes. The controller is implemented such
that the storageClassName, default, and description fields are managed
by the controller and cannot be changed by an admin. The admin can
change all other fields and the controller won't overwrite them.

Lastly, some re-factorization is done here to make the API groups in the
api-server more manageable.

Signed-off-by: Donnie Adams <donnie@acorn.io>