add: ImageAllowRules to restrict which images can run in a cluster/project
- ImageAllowRules

Currently we only support signature rules that require an image
to be signed by a specific set of keys and with a specific set of
annotations

Verifies Acorn images using `cosign` and if an image doesn't
passt the verifications specified by a set of ImageAllowRules, then it
won't be allowed to run

For now, we only have project-level resources. Cross-Project Resources
may be added later.

For now, the rules do not affect running images apart from updating a
status field. No rollbacks, etc. right now.

Signed-off-by: Thorsten Klein <tk@thklein.io>