Remove old defaults for volumes
Volumes size and access modes previously had built-in defaults of 10G
and readWriteOnce, respectively. However, the volume classes should now
be handling these defaults. Therefore, the old defaults should be
removed in favor of the volume class defaults.

Additionally, volume size was inconsistently parsed depending on whether
we were reading an Acornfile or JSON and whether the value was
interpreted to be an int or a string and how big the int was. This is
also fixed and tested here.

Signed-off-by: Donnie Adams <donnie@acorn.io>