Correctly wait for app updates
When doing an app update, the wait functionality didn't properly wait
for the update. It would see that the app was "ready" and stop. This was
incorrect because the app that was "ready" was the previous version.
Now, the wait functionality will ensure that it is waiting for the
correct version of the app before proceeding.

Additionally, there seems to be a flaky test. I am attempting to fix
that here, too.

Signed-off-by: Donnie Adams <donnie@acorn.io>