Handle default volume classes when cluster and project defaults exist
There was an error in logic when both a cluster and project volume class
defaults existed. Specifically, when the cluster default was
alphabetically smaller than the project default. This change addresses
this by ensuring that if there is a project volume class default then
the cluster volume class defaults are ignored.

Signed-off-by: Donnie Adams <donnie@acorn.io>