Change to HTTP readiness probe for build-server
The TCP probe tends to have issues, especially when running with
linkerd. This change switches the probe to an HTTP GET probe, which is
already enabled in the build-server.

A special note about this change: since this is changing the deployment
spec for the build-server, all currently running build-servers will be
restarted once this fix is pushed out to users.

Signed-off-by: Donnie Adams <donnie@acorn.io>