Add region API
Projects were previously not able to be updated. Since a project now
identifies which regions are supported, and these regions can be
added/removed, a user with the proper permissions can now update
projects.

Similar to other items, the project defines a default region. If an app
is created and does not specify a region, the default for the project is
used. An error is returned if a default region cannot be determined. A
user can override this default with the new --region run flag. Once a
region is defined for an app, it cannot change except for explicitly
using the region that is default for the project.

Signed-off-by: Donnie Adams <donnie@acorn.io>