Redesign Region API and add regioned objects
The Region API is overhauled here such that there is no longer any
RegionInstances. Additionally, a user will not be able to create a
region from the CLI, and a hard-coded "local" region will exist on
installation.

Finally, the various objects that must be specific to a region implement
an interface for detection purposes.

Signed-off-by: Donnie Adams <donnie@acorn.io>