Dont use local.on-acorn.io when acorn-dns disabled
This is a change of behavior to how we handle the acorn-dns install flag
being set to "disabled". Previously, we would still inject a
local.on-acorn.io endpoint if the user hadn't supplied a custom
clusterDomain. With this change, if acorn-dns is disabled, we won't
inject the local domain. This could potentially result in an app having
no endpoints, but that's now the expected behavior.

This fixes an issue we hit in our Hub clusters where we wanted to
control our app endpoints using the --publish syntax and then use
cert-manager to provision certs for the generated ingress, but because
the generated ingress would always have a local.on-acorn.io endpoint
(that would never correctly resolve in the ACME HTTP01 challenge), the
cert for the endpoint specified via --publish would never get a cert.

Signed-off-by: Craig Jellick <craig@acorn.io>