fix: custom Digest() func to error out on HEAD 404 (#1441)
When getting the digest for signature artifacts, crane.Digest() always
fell back to a GET request if HEAD failed, without checking the status
code. The custom function will just error out if HEAD returns a 404, as
it's common that the signature artifact is not present.

This should reduce the possibility of hitting e.g. DockerHub's rate
limit.

Signed-off-by: Thorsten Klein <tk@thklein.io>