Changes to prevent pull every upgrade interval for private images (#1398) (#1433)
* Changes to prevent pull every upgrade interval for private images (#1398)

Fixes Issue - 1398. The change allows findLatestTagForImageWithPattern to return if updated image is found or not and return what the new image is even when current tag to compare it with is empty. The changes to the function make it compare it with the current tag provided rather than the latest tag for that image.

Signed-off-by: Yashgiri Goswami <yashgirigoswami1@gmail.com>