Ensure app is removed before deleting other objects
There are situations when using `acorn rm -af` where the other objects
are deleted before the app is deleted. This causes things like generated
secrets to be regenerated and left around unexpectedly.

This change introduces a wait for app deletion that only happens when
the other objects are also being deleted. This ensures that the other
objects will be successfully removed.

Signed-off-by: Donnie Adams <donnie@acorn.io>