Handle edge-cases around auto-upgrade patterns
There were a few cases where auto-upgrade patterns failed to do the
expected thing.

1. When using an auto-upgrade pattern that did not match any existing
tags, the digest for existing tags would be pulled and used to start the
app. This is incorrect when no tag matched the pattern.
2. When using the auto-upgrade pattern of only '*', the matching would
fail to match any tag that was not "latest".
3. Using an auto-upgrade pattern ending with '**' implicitly expected
that the '**' would match something. So, if the pattern was 'v1.#**' and
the existing tags were 'v1.1-rc1' and 'v1.2', then the first tag would
match even though the second tag was "better" according to the '#'.

These issues are fixed here.

Co-authored-by: Yashgiri Goswami <yashgiri@acorn.io>
Signed-off-by: Donnie Adams <donnie@acorn.io>