Move region to spec of container replicas and volumes
The region was put on the status of container replicas and volumes
because the user doesn't directly set the region on these objects.
However, the user doesn't directly set anything on these objects; they
are generated from the app definition. Therefore, putting the region on
the spec makes sense and simplifies many things.

Signed-off-by: Donnie Adams <donnie@acorn.io>