Fix app and project region validation (#1542) (#1581)
* Fix app and project region validation

Previously, when a user removed a supported region from a project,
the apps in that region would be in various states. After these changes,
a user will be required to ensure that there are no apps running in a
region before removing that region as a supported region for the
project.

Additionally, if a user tried to change the region of a running app,
then the CLI command would silently ignore this. After these changes, an
error will be produced.

Signed-off-by: Donnie Adams <donnie@acorn.io>

---------

Signed-off-by: Donnie Adams <donnie@acorn.io>