Stop checking status for jobs that won't run
If a job is not expected to run for an event (like a delete job running
on a create event), then the status checker would wait for the job to
succeed before setting the app to a ready status. This will hang forever
because the job will never run. After these changes, the status checker
will ignore apps that are not expected to be created.

Signed-off-by: Donnie Adams <donnie@acorn.io>