Change behavior of update jobs and add stop jobs
The previous intended behavior of an update job was that the job would
run only when the job in the app was updated. This caused confusion
because others expected an update job to run whenever the app itself is
updated.

With these changes, these are the job types and their expected behavior:
- create jobs are run at least once, either when the app is created or
when the app is updated to have a create job
- update jobs are run whenever the app is updated, including start but
not including stop
- stop jobs are run whenever the app is stopped, not including deletion
- delete jobs are run when the app is deleted

Signed-off-by: Donnie Adams <donnie@acorn.io>