Reset statuses to ensure dependencies are handled correctly
If a job or a container has a dependency, then the controller will wait
correctly on creation. However, if the app is updated, then the
dependencies are already "ready" and Acorn was not distinguishing
between the old dependencies being ready and any new ones.

After this change, the dependency ready statuses are reset when the app
is updated to ensure that dependency reconciliation is handled properly.

Signed-off-by: Donnie Adams <donnie@acorn.io>