Ensure acme-challenge is available
To request a cert from let's encrypt for a custom domain, we have to
setup an acme HTTP01 challenge. This entails making a specific path
available with a specific value on the specified domain. To do so, we
setup an http server (that knows how to handle the acme protocol), a
service, and an ingress.

This addresses an issue where we were requesting the cert before the
server/servie/ingress combination was fully functional. It does so by
setting up a test path in the server and querying for a 200 response
before attempting to request the cert.

Signed-off-by: Craig Jellick <craig@acorn.io>