Stop processing tags on app creation
Based on which image was pulled when setting an image for an app, the
handler would append tags to the current image and (attempt to) remove
tags from other images. This causes confusion in the following scenario:
1. Create an app from an image tag.
2. Build an image with the same tag as above.
3. Depending on the tag, you may have two images with the same tag or
the image created in step one would remain tagged and the built image
would not.

The outcomes in step 3 are confusing. Therefore, after these changes,
the app handler will not try to "take" a tag from an existing image. The
idea is that if the desired tag exists when the app is run, then we
would have used that image. If the image that the app is using doesn't
have the tag, then a new image must have been built and tagged, which
"took over" the tag in question.

Signed-off-by: Donnie Adams <donnie@acorn.io>