Don't create or modify images after the image is set for an app
After the image for an app is set and tagged, a user still might want to
move tags around. Doing so would cause the controller to reconcile apps
and may double-tag or remove tags that the user just created.

After this change, the controller won't do anything with an image after
the image digest has been "observed." So, if a user moves tags around
via the CLI, then the controller won't touch the tags until the image
for the app explicitly changes.

This also reverts commit e96ba01a15bfb2044cf081f44c9a225c3b3b199b, which
was a previous attempt at addressing this issue.

Signed-off-by: Donnie Adams <donnie@acorn.io>