Put project region information on status
Previously, the default region on status was only set if the user didn't
specify a default region for the project. As soon as the user specified
a default region, the field on the project's status would be cleared.

This behavior meant that the controllers and the UI would have to check
the spec and the status to determine default and supported regions. Not
ideal.

After this change, the project's status will be populated with the
default and supported regions regardless of whether the user set them.
If the user set them on the spec, then these values will simply be
copied to the status. The goal is that the controllers and the UI now
have one place to check for a default and supported regions: the status
of the project.

Signed-off-by: Donnie Adams <donnie@acorn.io>