Add ignore-cleanup flag to rm command
The ignore-cleanup flag will allow a user to forego any delete jobs that
an app might have. The idea here is to create an escape hatch for a user
who may have misconfigured their delete job.

This functionality only allows a user to ignore the delete job. If the
app is stuck deleting for another reason, then this is out of the user's
responsibility.

Secondly, it is an error to call the IgnoreCleanup subresource on an app
that is not deleting.

Signed-off-by: Donnie Adams <donnie@acorn.io>