Use AppInstance instead of App for app subresouces that update
The subresource calls for Apps in Hub get forwarded to the workload
cluster. Updating the App (instead of the AppInstance) in the workload
cluster will cause the App to go through validation. We don't want this
to happen in the workload cluster because some objects, like
compute classes, are slightly modified when they are synced up.
Validating at the account-api-server level works, but not in the
workload cluster.

This change will use the AppInstance for these subresources to avoid
this error.

Signed-off-by: Donnie Adams <donnie@acorn.io>