Fix stack overflow with circular service dependencies
If an AppInstance CR for a service acorn was yet to be created because
of a pending secret or other resource it was possible that a circular
reference could be created and cause a stack overflow.  Circular references
are now checked and for this situation where it could happen temporarily
we look for that specific situation.

Signed-off-by: Darren Shepherd <darren@acorn.io>