Create the CRDs before creating the router
When a router is created, the RESTMapper is created from the rest config
at the same time. If the CRDs don't exist by the time the RESTMapper is
created, then starting the router will fail to find the resource for the
GVKs. This will cause the acorn-controller to restart.

This change creates the CRDs before creating the router, ensuring that
the RESTMapper has all needed data.

Signed-off-by: Donnie Adams <donnie@acorn.io>