Add project instances
There are a few situations where we need a buffer between a project
being deleted and the corresponding namespace being deleted. One such
scenario is ensuring that delete jobs run when a project is deleted.

This change adds a project instance, which is this buffer. Using the
project instance, we can ensure that all apps are successfully cleaned
up before the namespace is deleted. This should also give the user the
ability to use the ignore-cleanup functionality when deleting a project.

Changes are also made to the integration tests to use project instances
instead of namespaces.

Signed-off-by: Donnie Adams <donnie@acorn.io>