Switch to creating a single wildcard record per install
This commit removes the logic where each Ingress created by Apps would
result in a record being registered through Acorn DNS. Now, we create a
single "acorn-ingress" Ingress object that gets processed like the old
App Ingresses were. This allows us to create a single wildcard record
for the entire install, which is much more efficient.

In addition to this change, this also switches to creating and using a
mocked DNS client for testing.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>