Wait for all parent apps to be deleted before deleting children
When a project is deleted, we clean up all apps. The problem with this
current implementation is that all apps will be deleted, regardless of
whether it was a child app. This leads to child apps have issues
deleting.

This change will first ensure all "parent acorns" are deleted first,
giving the child acorns time to delete. After all such acorns are
deleted, then the remaining child acorns are also deleted.

Signed-off-by: Donnie Adams <donnie@acorn.io>