Update auto-upgrade logic to have less assumptions
The original auto-upgrade assumed that setting an app to auto-upgrade
could be done by:
1. Setting AutoUpgrade to true
2. Having a tag is an auto-upgrade pattern
3. Setting an interval
4. Setting NotifyUpgrade to true

Now the only two ways to set an app to auto-upgrade are:
1. Setting AutoUpgrade to true
2. Having a tag is an auto-upgrade pattern

To patch up the UX such that it is the same as before from the CLI's
perspective there is new code that sets `AutoUpgrade` to true if it is
implied. This effectually mantains the same UX as before with the backend
behaving in a more explicit manner.

The unit test for the auto-upgrade daemon has been updated to reflect
this new logic.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>