Add the ability to set the log level at startup
Previously, in order to change the log level, one must exec into each
container and run a command. This is nice for existing containers, but
can be tedious if trying to run containers from the start with a log
level different from "info."

This change adds the ability to set an environment variable and start
the container with a desired log level.

Signed-off-by: Donnie Adams <donnie@acorn.io>