Use context.AfterFunc instead of goroutine approach
Note that the stop function returned by context.AfterFunc doesn't run
the function that is passed to context.AfterFunc. It disassociates the
function from the context (i.e. releases the resources). Calling stop
after the context is cancelled is valid, so stop should be called to
ensure that the resources are released.

Signed-off-by: Donnie Adams <donnie@acorn.io>