Fix dependency checks for auto-upgrade
During an auto-upgrade, the generation of an Acorn does not change.
Therefore, the checks for "is the whole acorn up to date" don't stop
premature upgrades for dependent objects.

This change checks images and changes the order of the status checks to
ensure that dependent objects do not upgrade before their dependencies
are ready.

Signed-off-by: Donnie Adams <donnie@acorn.io>