Change auth url for logging into manager
By going to /auth/login every time, we are forcing the browser to do a
full logout/login. /auth/cli will only do a login if the token is
expired, so it's a little friendlier UX (and let's us test the switch
from beta.acorn.io to acorn.io more easily).

Signed-off-by: Craig Jellick <craig@acorn.io>