Set dev session region if not set
When starting a dev session on a new app without setting a region on the
app, an error might crop up complaining about regions. This does not
happen if a region is explicitly set on the app.

Therefore, since users don't explicitly create dev session objects,
setting the region for them when the dev session region is unset is
reasonable.

Signed-off-by: Donnie Adams <donnie@acorn.io>