Only set Status.Defaults.Volumes once
This also includes a migration step away from Status.Defaults.VolumeSize. It
removes the logic that originally set the VolumeSize field and instead hands
that logic off to the `Status.Defaults.Volumes` map field. If VolumeSize was
previously set, then we will set that in the Volumes map field as the default.
This prevents any issues from occuring with existing volumes when the default
VolumeClass is updated.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>