Add pre-stop hooks for containers that expose ports
This pre-stop hook will simply wait for 5 seconds. In order to achieve
this in the most general way, a configmap is mounted as a volume that
has a very simply binary that sleeps for 5 seconds. It is done this way
because this should work for any container images, including scratch.

Wait is not applied for stateful containers, jobs, or containers with
one replica.

Signed-off-by: Donnie Adams <donnie@acorn.io>