Stop rerunning create jobs on non-generation bumping updates
If an app is updated in such a way that the generation is not bumped
(e.g. an annotation or label is add/removed/modified), then any job that
is create-only would rerun if the generation of the app is 1. This
change ensures that create-only jobs will not be rerun in this instance.

Signed-off-by: Donnie Adams <donnie@acorn.io>