Ensure that nginx does DNS lookups
When using static hostnames, nginx will lookup the DNS for these
hostnames on startup/config reload and won't do it again until the
container is restarted or the config is reloaded. This means that if the
service IPs change, then nginx won't have the correct IPs until one of
these things happen.

This change uses a hack to ensure that nginx will look up for the DNS
for services at least every 15 seconds. This ensures that if the service
IPs change for some reason, then nginx will get them.

Signed-off-by: Donnie Adams <donnie@acorn.io>