More reliably determine when linked/aliased services are ready
A linked or aliased service could be in a different namespace that the
running app. This means that the hash check that is done is erroneous
because the service instance that is inspected could be from an entirely
different app. This change better detects with by moving the namespace
check.

Additionally, a few small cleanup changes are included here.

Signed-off-by: Donnie Adams <donnie@acorn.io>