fix: add CPU field to resolvedOfferings and change QuotaRequests to use resolvedOfferings
This change makes it so that we always quota against the resolvedOfferings instead
of the scheduling for the app. This is because the scheduling for the app is always
accurate to what is actually running, but the resolvedOfferings is what the user
requested.

This change also adds a CPU field to the resolvedOfferings so that we can quota
against the CPU that the user requested from one place instead of having to calculate
it from the cpuScaler.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>